Frequency Sweep
Added a sweep across the frequency spectrum. Demonstrates the circuit
can operate across the entire FM spectrum. Capacitance was solved for
analytically as a function of frequency. Capacitance should be modulated
rather than inductance, as to not affect the bandwidth. Conversely,
resistance should be modulated for bandwidth rather than inductance, as
it does not affect the tuned frequency.